[Refactor]#1 Networkable 삭제를 통한 의존성 주입
